Many things

+++ Server/Server.py
@@ -1,29 +1,65 @@
 import json
 import urllib
 from http.server import BaseHTTPRequestHandler, HTTPServer, SimpleHTTPRequestHandler
-import time
+import time,calendar
 
 import os
 import qrcode
 from Data.entities import worlds
+from functools import reduce
 
 hostName = "0.0.0.0"
 hostPort = 8080
-
-
+people_in_stores = {
+"Castro":set([(1,1),(2,2),(3,3)]),
+"FOX HOME":set([(1,1),(2,2),(3,3),(4,4)]),
+"Avazi":set([(1,1),(2,2)])
+}
+category_to_time = {
+    "clothing":5,
+    "food":3,
+    "drink":1,
+    "coffee":1,
+    "book":21,
+    "wheels":45,
+    "medicine":14,
+    "locks":18
+}
 # generate qr
 def getQR(id):
     print('getQR ({})'.format(id))
-    img = qrcode.make(json.dumps({'id': id, 'creation_date': time.time()}))
+    img = qrcode.make(json.dumps({'id': id, 'creation_date': get_now()}))
     url = '{}.png'.format(id)
     img.save(url)
     print(url)
     return url
 
+def get_now():
+    return calendar.timegm(time.gmtime())
+def scannedQR(store, id, creation_date):
+    print('scannedQR ({}, {}, {})'.format(store, id, creation_date))
+    categories = None
+    for world in worlds.values():
+        for s in world.get_services().values():
+            if store == s.get_name():
+                categories = s.categories
+    l = [category_to_time[cat] for cat in categories]
+    activity_time = reduce(lambda x, y: x + y, l) / len(l)
+    # for (id,creation_date) in people_in_stores[store]:
+    #     if creation_date + activity_time < get_now():
+    #         people_in_stores[store].discard((id,creation_date))
+    #         print("removed by time")
+    if store not in people_in_stores:
+        people_in_stores[store] = set()
+    if (id,creation_date) in people_in_stores[store]:
+        people_in_stores[store].discard((id,creation_date))
+        os.remove('{}.png'.format(id))
+        print("removed")
+    else:
+        people_in_stores[store].add((id,creation_date))
+        print("Added")
 
-def scannedQR(id, creation_date):
-    print('scannedQR ({}, {})'.format(id, creation_date))
-    os.remove('{}.png'.format(id))
+    # print(people_in_stores[store])
     return 'scanned successfully'
 
 
@@ -32,15 +68,21 @@ def clientLeft(params):
 
 
 def getWorlds():
-    print('getWorlds')
+    print("getWorlds")
     return str([w.get_name() for w in worlds.values()])
 
 
 def getWorld(id):
     print('getWorld ({})'.format(id))
     services = list(worlds[id].get_services().values())
-    ser = [json.dumps(s.__dict__()) for s in services]
-    return str(ser)
+    services = [s.__dict__() for s in services]
+    for s in services:
+        if s["name"] not in people_in_stores:
+            people_in_stores[s["name"]] = set()
+        s["crowd"] = len(people_in_stores[s["name"]])
+        s["status"] = int(min(len(people_in_stores[s["name"]])/2,2))
+    ser = [json.dumps(s) for s in services]
+    return str(ser).replace("\'","")
 
 
 def register(id):
@@ -83,6 +125,7 @@ class MyServer(SimpleHTTPRequestHandler):
                 return super().do_GET()
             params = parse_params(params_string)
             self.send_response(200)
+            self.send_header('Access-Control-Allow-Origin', '*')
             self.end_headers()
             print(params)
             self.wfile.write(functions[req](**params).encode("utf-8"))
